Rename service and base service files
