+++ docs/new-algorithm-service.md
@@ -17,7 +17,7 @@ from pkg.apis.manager.v1alpha3.python import api_pb2
 from pkg.apis.manager.v1alpha3.python import api_pb2_grpc
 from pkg.suggestion.v1alpha3.internal.search_space import HyperParameter, HyperParameterSearchSpace
 from pkg.suggestion.v1alpha3.internal.trial import Trial, Assignment
-from pkg.suggestion.v1alpha3.hyperopt.base_hyperopt_service import BaseHyperoptService
+from pkg.suggestion.v1alpha3.hyperopt.base_service import BaseHyperoptService
 from pkg.suggestion.v1alpha3.base_health_service import HealthServicer
 
 
@@ -122,7 +122,7 @@ import unittest
 from pkg.apis.manager.v1alpha3.python import api_pb2_grpc
 from pkg.apis.manager.v1alpha3.python import api_pb2
 
-from pkg.suggestion.v1alpha3.hyperopt_service import HyperoptService
+from pkg.suggestion.v1alpha3.hyperopt.service import HyperoptService
 
 class TestHyperopt(unittest.TestCase):
     def setUp(self):
